Moved timer implementation to separate module.
